requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
idna==2.10
Markdown==3.2.2
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
